+++ packages/manager/modules/nasha/src/dashboard/translations/Messages_fr_FR.json
